adapters/llamaindex/pyproject.toml
[project]
name = "voidly-pay-llamaindex"
version = "1.0.0"
description = "LlamaIndex tools for the Voidly Pay agent marketplace."
authors = [{ name = "Voidly Research", email = "research@voidly.ai" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = ["voidly-pay>=1.0.0", "llama-index-core>=0.10.0"]

[project.urls]
Homepage = "https://voidly.ai/pay"
Repository = "https://github.com/voidly-ai/voidly-pay/tree/main/adapters/llamaindex"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["voidly_pay_llamaindex*"]
